+++ AUTHORS.rst
@@ -3,3 +3,5 @@
 * Alireza Savand <alireza.savand@gmail.com>
 * Martin Burchell <martinb@aptivate.org>
 * Daniel Lawrence <dannyla@linux.com>
+* Fabio C. Barrionuevo da Luz <bnafta@gmail.com>
+
