--- run_tests.py
@@ -1,43 +1,0 @@
-import os
-import sys
-
-import django
-from django.conf import settings
-
-DIR_NAME = os.path.dirname(__file__)
-settings.configure(
-    DEBUG=True,
-    DATABASES={
-        'default': {
-            'ENGINE': 'django.db.backends.sqlite3',
-        }
-    },
-    INSTALLED_APPS=(
-        'django.contrib.admin',
-        'django.contrib.auth',
-        'django.contrib.contenttypes',
-        'django.contrib.sessions',
-        'cuser',
-    ),
-    ROOT_URLCONF='tests.CuserTestCase.urls',
-    MIDDLEWARE_CLASSES = [
-        'django.middleware.security.SecurityMiddleware',
-        'django.contrib.sessions.middleware.SessionMiddleware',
-        'django.middleware.common.CommonMiddleware',
-        'django.middleware.csrf.CsrfViewMiddleware',
-        'django.contrib.auth.middleware.AuthenticationMiddleware',
-        'django.contrib.auth.middleware.SessionAuthenticationMiddleware',
-        'django.contrib.messages.middleware.MessageMiddleware',
-        'django.middleware.clickjacking.XFrameOptionsMiddleware',
-        'cuser.middleware.CuserMiddleware',
-    ]
-)
-
-from django.test.runner import DiscoverRunner
-django.setup()
-
-test_runner = DiscoverRunner(verbosity=2)
-failures = test_runner.run_tests(['cuser', ])
-
-if failures:
-    sys.exit(failures)
